+++ isofit/radiative_transfer/sRTMnet.py
@@ -33,9 +33,9 @@ from isofit.configs.sections.radiative_transfer_config import (
 )
 from isofit.core.common import resample_spectrum
 from isofit.core.sunposition import sunpos
+from isofit.radiative_transfer import luts
 from isofit.radiative_transfer.radiative_transfer_engine import RadiativeTransferEngine
 from isofit.radiative_transfer.six_s import SixSRT
-from isofit.utils import luts
 
 Logger = logging.getLogger(__file__)
